+++ index.js
@@ -34,9 +34,14 @@ class LgWebOsPlatform {
 					return;
 				}
 
-				//debug config
+				//log config
 				const enableDebugMode = device.enableDebugMode || false;
-				const debug = enableDebugMode ? log.info(`Device: ${host} ${deviceName}, Did finish launching.`) : false;
+				const disableLogDeviceInfo = device.disableLogDeviceInfo || false;
+				const disableLogInfo = device.disableLogInfo || false;
+				const disableLogSuccess = device.disableLogSuccess || false;
+				const disableLogWarn = device.disableLogWarn || false;
+				const disableLogError = device.disableLogError || false;
+				const debug = !enableDebugMode ? false : log.info(`Device: ${host} ${deviceName}, did finish launching.`);
 				const config = {
 					...device,
 					mqtt: {
@@ -44,10 +49,11 @@ class LgWebOsPlatform {
 						passwd: 'removed'
 					}
 				};
-				const debug1 = enableDebugMode ? log.info(`Device: ${host} ${deviceName}, Config: ${JSON.stringify(config, null, 2)}.`) : false;
+				const debug1 = !enableDebugMode ? false : log.info(`Device: ${host} ${deviceName}, Config: ${JSON.stringify(config, null, 2)}.`);
 
 				//check files exists, if not then create it
 				const postFix = host.split('.').join('');
+				const certFile = `${prefDir}/cert_${postFix}`;
 				const keyFile = `${prefDir}/key_${postFix}`;
 				const devInfoFile = `${prefDir}/devInfo_${postFix}`;
 				const inputsFile = `${prefDir}/inputs_${postFix}`;
@@ -57,6 +63,7 @@ class LgWebOsPlatform {
 
 				try {
 					const files = [
+						certFile,
 						keyFile,
 						devInfoFile,
 						inputsFile,
@@ -71,34 +78,34 @@ class LgWebOsPlatform {
 						}
 					});
 				} catch (error) {
-					log.error(`Device: ${host} ${deviceName}, Prepare files error: ${error}.`);
+					const emitLog = disableLogError ? false : log.error(`Device: ${host} ${deviceName}, Prepare files error: ${error}.`);
 					return;
 				}
 
 				//webos device
 				try {
-					const lgWebOsDevice = new LgWebOsDevice(api, device, keyFile, devInfoFile, inputsFile, channelsFile, inputsNamesFile, inputsTargetVisibilityFile);
+					const lgWebOsDevice = new LgWebOsDevice(api, device, certFile, keyFile, devInfoFile, inputsFile, channelsFile, inputsNamesFile, inputsTargetVisibilityFile);
 					lgWebOsDevice.on('publishAccessory', (accessory) => {
 						api.publishExternalAccessories(PluginName, [accessory]);
-						log.success(`Device: ${host} ${deviceName}, Published as external accessory.`);
+						const emitLog = disableLogSuccess ? false : log.success(`Device: ${host} ${deviceName}, Published as external accessory.`);
 					})
 						.on('devInfo', (devInfo) => {
-							log.info(devInfo);
+							const emitLog = disableLogDeviceInfo ? false : log.info(devInfo);
 						})
-						.on('success', (message) => {
-							log.success(`Device: ${host} ${deviceName}, ${message}.`);
+						.on('success', (success) => {
+							const emitLog = disableLogSuccess ? false : log.success(`Device: ${host} ${deviceName}, ${success}.`);
 						})
-						.on('message', (message) => {
-							log.info(`Device: ${host} ${deviceName}, ${message}.`);
+						.on('info', (info) => {
+							const emitLog = disableLogInfo ? false : log.info(`Device: ${host} ${deviceName}, ${info}.`);
 						})
 						.on('debug', (debug) => {
-							log.info(`Device: ${host} ${deviceName}, debug: ${debug}.`);
+							const emitLog = !enableDebugMode ? false : log.info(`Device: ${host} ${deviceName}, debug: ${debug}.`);
 						})
 						.on('warn', (warn) => {
-							log.warn(`Device: ${host} ${deviceName}, ${warn}.`);
+							const lemitLog = disableLogWarn ? false : log.warn(`Device: ${host} ${deviceName}, ${warn}.`);
 						})
 						.on('error', (error) => {
-							log.error(`Device: ${host} ${deviceName}, ${error}.`);
+							const emitLog = disableLogError ? false : log.error(`Device: ${host} ${deviceName}, ${error}.`);
 						});
 
 					//create impulse generator
@@ -108,10 +115,10 @@ class LgWebOsPlatform {
 							const startDone = await lgWebOsDevice.start();
 							const stopImpulseGenerator = startDone ? await impulseGenerator.stop() : false;
 						} catch (error) {
-							log.error(`Device: ${host} ${deviceName}, ${error}, trying again.`);
+							const emitLog = disableLogError ? false : log.error(`Device: ${host} ${deviceName}, ${error}, trying again.`);
 						};
 					}).on('state', (state) => {
-						const debug = enableDebugMode ? state ? log.info(`Device: ${host} ${deviceName}, Start impulse generator started.`) : log.info(`Device: ${host} ${deviceName}, Start impulse generator stopped.`) : false;
+						const emitLog = !enableDebugMode ? false : state ? log.info(`Device: ${host} ${deviceName}, Start impulse generator started.`) : log.info(`Device: ${host} ${deviceName}, Start impulse generator stopped.`);
 					});
 
 					//start impulse generator